Revert "Interfacce"
