+++ dao/UtenteDaoImpl.java
@@ -44,12 +44,8 @@ public class UtenteDaoImpl extends DaoImpl<Utente>{
     }
 
     @Override
-    public Utente addItem(Utente item) throws ExecutionException, InterruptedException {
-        DocumentReference documentReference = firestore.collection(collectionName).document(item.generateId());
-        ApiFuture<WriteResult> writeResult = documentReference.set(item, SetOptions.merge());
-
-        DocumentSnapshot documentSnapshot = documentReference.get().get();
-        return getItem(documentSnapshot.getId());
+    public Utente addItem(Utente item) {
+        return null;
     }
 
     @Override
