+++ views/InsertPersonaleContratto.fxml
@@ -1,0 +1,74 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import com.jfoenix.controls.JFXButton?>
+<?import com.jfoenix.controls.JFXComboBox?>
+<?import com.jfoenix.controls.JFXDatePicker?>
+<?import com.jfoenix.controls.JFXTextField?>
+<?import javafx.scene.control.Tooltip?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.text.Font?>
+<?import javafx.scene.text.Text?>
+
+<VBox prefHeight="519.0" prefWidth="848.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="elaborato_ingegneriaSW.controllers.InsertPersonaleContrattoController">
+  <children>
+    <AnchorPane maxHeight="-1.0" maxWidth="-1.0" prefHeight="539.0" prefWidth="790.0" style="-fx-background-color: #faf0f0; -fx-effect: dropshadow(three-pass-box, rgba(0, 0, 0, 1.2), 5, 0, 0, 0);" VBox.vgrow="ALWAYS">
+         <children>
+            <JFXButton layoutX="750.0" layoutY="444.0" opacity="0.0" prefHeight="44.0" prefWidth="84.0">
+               <tooltip>
+                  <Tooltip text="TODO: aggiungere qualcosa" />
+               </tooltip>
+            </JFXButton>
+            <Text layoutX="73.0" layoutY="59.0" strokeType="OUTSIDE" strokeWidth="0.0" text="INSERIMENTO PERSONALE  A CONTRATTO" textAlignment="CENTER" wrappingWidth="701.5359497070312">
+               <font>
+                  <Font name="System Bold" size="33.0" />
+               </font>
+            </Text>
+            <Text layoutX="116.0" layoutY="137.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Nome">
+               <font>
+                  <Font name="Open Sans" size="20.0" />
+               </font>
+            </Text>
+            <Text layoutX="116.0" layoutY="194.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Cognome">
+               <font>
+                  <Font name="Open Sans" size="20.0" />
+               </font>
+            </Text>
+            <Text layoutX="116.0" layoutY="254.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Data di nascita">
+               <font>
+                  <Font name="Open Sans" size="20.0" />
+               </font>
+            </Text>
+            <Text layoutX="116.0" layoutY="378.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Comune associato">
+               <font>
+                  <Font name="Open Sans" size="20.0" />
+               </font>
+            </Text>
+            <JFXComboBox fx:id="comuneAssociatoComboBox" focusColor="#7c2e2e" layoutX="347.0" layoutY="357.0" prefHeight="26.0" prefWidth="340.0" />
+            <JFXButton fx:id="insertPersonaleButton" buttonType="RAISED" defaultButton="true" layoutX="331.0" layoutY="434.0" onAction="#insertPersonaleAction" prefHeight="40.0" prefWidth="198.0" ripplerFill="WHITE" style="-fx-background-color: #7b241c; -fx-background-radius: 15pt; -fx-border-radius: 15pt; -fx-border-color: black;" text="INSERISCI PERSONA" textFill="WHITE">
+               <font>
+                  <Font name="Open Sans Semibold" size="15.0" />
+               </font>
+            </JFXButton>
+            <JFXTextField fx:id="nomeTextField" focusColor="#7c2e2e" layoutX="347.0" layoutY="116.0" prefHeight="26.0" prefWidth="340.0">
+               <font>
+                  <Font name="Open Sans" size="18.0" />
+               </font></JFXTextField>
+            <JFXTextField fx:id="cfTextField" focusColor="#7c2e2e" layoutX="347.0" layoutY="290.0" prefHeight="26.0" prefWidth="340.0">
+               <font>
+                  <Font name="Open Sans" size="18.0" />
+               </font></JFXTextField>
+            <JFXTextField fx:id="cognomeTextField" focusColor="#7c2e2e" layoutX="347.0" layoutY="173.0" prefHeight="26.0" prefWidth="340.0">
+               <font>
+                  <Font name="Open Sans" size="18.0" />
+               </font></JFXTextField>
+            <Text layoutX="116.0" layoutY="318.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Codice Fiscale">
+               <font>
+                  <Font name="Open Sans" size="20.0" />
+               </font>
+            </Text>
+            <JFXDatePicker fx:id="dataNascitaDatePicker" defaultColor="#7c2e2e" layoutX="343.0" layoutY="245.0" prefHeight="28.0" prefWidth="346.0" />
+         </children>
+    </AnchorPane>
+  </children>
+</VBox>
