--- views/InsertUtente.fxml
@@ -1,98 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import com.jfoenix.controls.JFXButton?>
-<?import com.jfoenix.controls.JFXComboBox?>
-<?import com.jfoenix.controls.JFXTextField?>
-<?import javafx.scene.control.CheckBox?>
-<?import javafx.scene.control.Tooltip?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.text.Font?>
-<?import javafx.scene.text.Text?>
-
-<VBox prefHeight="519.0" prefWidth="848.0" style="-fx-background-color: trasparent;" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="elaborato_ingegneriaSW.controllers.InsertUtenteController">
-  <children>
-      <AnchorPane maxHeight="-1.0" maxWidth="-1.0" prefHeight="539.0" prefWidth="790.0" style="-fx-background-color: #f4f6f6; -fx-border-radius: 20px; -fx-background-radius: 20px; -fx-effect: dropshadow(three-pass-box, rgba(0, 0, 0, 1.2), 5, 0, 0, 0);" VBox.vgrow="ALWAYS">
-         <children>
-            <JFXButton layoutX="750.0" layoutY="444.0" opacity="0.0" prefHeight="44.0" prefWidth="84.0">
-               <tooltip>
-                  <Tooltip text="TODO: aggiungere qualcosa" />
-               </tooltip>
-            </JFXButton>
-            <Text layoutX="73.0" layoutY="59.0" strokeType="OUTSIDE" strokeWidth="0.0" text="INSERIMENTO UTENTE" textAlignment="CENTER" wrappingWidth="701.5359497070312">
-               <font>
-                  <Font name="System Bold" size="33.0" />
-               </font>
-            </Text>
-            <Text layoutX="28.0" layoutY="131.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Nome">
-               <font>
-                  <Font name="Open Sans" size="20.0" />
-               </font>
-            </Text>
-            <Text layoutX="424.0" layoutY="131.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Cognome">
-               <font>
-                  <Font name="Open Sans" size="20.0" />
-               </font>
-            </Text>
-            <Text layoutX="45.0" layoutY="322.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Comuni">
-               <font>
-                  <Font name="Open Sans" size="20.0" />
-               </font>
-            </Text>
-            <JFXButton fx:id="insertUtenteButton" buttonType="RAISED" defaultButton="true" layoutX="331.0" layoutY="434.0" onAction="#insertUtenteAction" prefHeight="40.0" prefWidth="198.0" ripplerFill="WHITE" style="-fx-background-color: #7b241c; -fx-background-radius: 15pt; -fx-border-radius: 15pt; -fx-border-color: black;" text="INSERISCI UTENTE" textFill="WHITE">
-               <font>
-                  <Font name="Open Sans Semibold" size="15.0" />
-               </font>
-            </JFXButton>
-            <JFXTextField fx:id="nomeTextField" focusColor="#7c2e2e" layoutX="102.0" layoutY="109.0" prefHeight="38.0" prefWidth="268.0">
-               <font>
-                  <Font name="Open Sans" size="18.0" />
-               </font>
-            </JFXTextField>
-            <JFXTextField fx:id="cfTextField" focusColor="#7c2e2e" layoutX="175.0" layoutY="224.0" prefHeight="38.0" prefWidth="254.0">
-               <font>
-                  <Font name="Open Sans" size="18.0" />
-               </font>
-            </JFXTextField>
-            <JFXTextField fx:id="cognomeTextField" focusColor="#7c2e2e" layoutX="529.0" layoutY="110.0" prefHeight="28.0" prefWidth="253.0">
-               <font>
-                  <Font name="Open Sans" size="18.0" />
-               </font>
-            </JFXTextField>
-            <Text layoutX="28.0" layoutY="253.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Codice Fiscale">
-               <font>
-                  <Font name="Open Sans" size="20.0" />
-               </font>
-            </Text>
-            <Text layoutX="25.0" layoutY="193.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Username">
-               <font>
-                  <Font name="Open Sans" size="20.0" />
-               </font>
-            </Text>
-            <JFXTextField fx:id="usernameTextField" focusColor="#7c2e2e" layoutX="141.0" layoutY="170.0" prefHeight="38.0" prefWidth="233.0">
-               <font>
-                  <Font name="Open Sans" size="18.0" />
-               </font>
-            </JFXTextField>
-            <Text layoutX="423.0" layoutY="193.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Password">
-               <font>
-                  <Font name="Open Sans" size="20.0" />
-               </font>
-            </Text>
-            <JFXTextField fx:id="passwordTextField" focusColor="#7c2e2e" layoutX="528.0" layoutY="170.0" prefHeight="28.0" prefWidth="254.0">
-               <font>
-                  <Font name="Open Sans" size="18.0" />
-               </font>
-            </JFXTextField>
-            <CheckBox layoutX="160.0" layoutY="302.0" mnemonicParsing="false" text="CheckBox" />
-            <CheckBox layoutX="160.0" layoutY="337.0" mnemonicParsing="false" text="CheckBox" />
-            <Text layoutX="459.0" layoutY="253.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Ruolo">
-               <font>
-                  <Font name="Open Sans" size="20.0" />
-               </font>
-            </Text>
-            <JFXComboBox fx:id="ruoloComboBox" layoutX="528.0" layoutY="231.0" prefHeight="28.0" prefWidth="254.0" />
-         </children>
-      </AnchorPane>
-  </children>
-</VBox>
